feat(desk): use document layout from Components API
