+++ packages/sanity/src/desk/panes/document/DocumentPaneProvider.tsx
@@ -1,39 +1,30 @@
-import React, {memo, useCallback, useEffect, useMemo, useRef, useState} from 'react'
+import {memo, useCallback, useEffect, useMemo, useRef, useState} from 'react'
 import type {ObjectSchemaType, Path, SanityDocument, SanityDocumentLike} from '@sanity/types'
 import {omit} from 'lodash'
 import {useToast} from '@sanity/ui'
 import {fromString as pathFromString, resolveKeyedPath} from '@sanity/util/paths'
-import isHotkey from 'is-hotkey'
 import {isActionEnabled} from '@sanity/schema/_internal'
 import {usePaneRouter} from '../../components'
 import type {PaneMenuItem} from '../../types'
 import {useDeskTool} from '../../useDeskTool'
 import {structureLocaleNamespace} from '../../i18n'
-import {CommentsProvider, CommentsSelectedPathProvider, useCommentsEnabled} from '../../comments'
 import {DocumentPaneContext, type DocumentPaneContextValue} from './DocumentPaneContext'
-import {getMenuItems} from './menuItems'
 import type {DocumentPaneProviderProps} from './types'
 import {usePreviewUrl} from './usePreviewUrl'
 import {getInitialValueTemplateOpts} from './getInitialValueTemplateOpts'
 import {
-  COMMENTS_INSPECTOR_NAME,
   DEFAULT_MENU_ITEM_GROUPS,
   EMPTY_PARAMS,
   HISTORY_INSPECTOR_NAME,
   INSPECT_ACTION_PREFIX,
 } from './constants'
-import {DocumentInspectorMenuItemsResolver} from './DocumentInspectorMenuItemsResolver'
 import {
   type DocumentFieldAction,
-  type DocumentFieldActionNode,
   type DocumentInspector,
-  type DocumentInspectorMenuItem,
   type DocumentPresence,
   type PatchEvent,
   type StateTree,
   EMPTY_ARRAY,
-  FieldActionsProvider,
-  FieldActionsResolver,
   getDraftId,
   getExpandOperations,
   getPublishedId,
@@ -113,8 +104,6 @@ export const DocumentPaneProvider = memo((props: DocumentPaneProviderProps) => {
   const schemaType = schema.get(documentType) as ObjectSchemaType | undefined
   const value: SanityDocumentLike = editState?.draft || editState?.published || initialValue.value
   const [isDeleting, setIsDeleting] = useState(false)
-
-  const [inspectorMenuItems, setInspectorMenuItems] = useState<DocumentInspectorMenuItem[]>([])
 
   // Resolve document actions
   const actions = useMemo(
@@ -220,21 +209,8 @@ export const DocumentPaneProvider = memo((props: DocumentPaneProviderProps) => {
 
   const changesOpen = currentInspector?.name === HISTORY_INSPECTOR_NAME
 
-  const hasValue = Boolean(value)
   const {t} = useTranslation(structureLocaleNamespace)
-  const menuItems = useMemo(
-    () =>
-      getMenuItems({
-        currentInspector,
-        features,
-        hasValue,
-        inspectorMenuItems,
-        inspectors,
-        previewUrl,
-        t,
-      }),
-    [currentInspector, features, hasValue, inspectorMenuItems, inspectors, previewUrl, t],
-  )
+
   const inspectOpen = params.inspect === 'on'
   const compareValue: Partial<SanityDocument> | null = changesOpen
     ? sinceAttributes
@@ -473,22 +449,6 @@ export const DocumentPaneProvider = memo((props: DocumentPaneProviderProps) => {
     ],
   )
 
-  const handleKeyUp = useCallback(
-    (event: React.KeyboardEvent<HTMLDivElement>) => {
-      for (const item of menuItems) {
-        if (item.shortcut) {
-          if (isHotkey(item.shortcut, event)) {
-            event.preventDefault()
-            event.stopPropagation()
-            handleMenuAction(item)
-            return
-          }
-        }
-      }
-    },
-    [handleMenuAction, menuItems],
-  )
-
   const handleLegacyInspectClose = useCallback(
     () => toggleLegacyInspect(false),
     [toggleLegacyInspect],
@@ -616,7 +576,6 @@ export const DocumentPaneProvider = memo((props: DocumentPaneProviderProps) => {
     focusPath,
     inspector: currentInspector || null,
     inspectors,
-    menuItems,
     onBlur: handleBlur,
     onChange: handleChange,
     onFocus: handleFocus,
@@ -624,7 +583,6 @@ export const DocumentPaneProvider = memo((props: DocumentPaneProviderProps) => {
     onHistoryClose: handleHistoryClose,
     onHistoryOpen: handleHistoryOpen,
     onInspectClose: handleLegacyInspectClose,
-    onKeyUp: handleKeyUp,
     onMenuAction: handleMenuAction,
     onPaneClose: handlePaneClose,
     onPaneSplit: handlePaneSplit,
@@ -699,66 +657,8 @@ export const DocumentPaneProvider = memo((props: DocumentPaneProviderProps) => {
     return undefined
   }, [params, documentId, onFocusPath, setOpenPath, ready, paneRouter])
 
-  const [rootFieldActionNodes, setRootFieldActionNodes] = useState<DocumentFieldActionNode[]>([])
-
-  const commentsEnabled = useCommentsEnabled()
-
-  const handleOpenCommentsInspector = useCallback(() => {
-    if (currentInspector?.name === COMMENTS_INSPECTOR_NAME) return
-
-    openInspector(COMMENTS_INSPECTOR_NAME)
-  }, [currentInspector?.name, openInspector])
-
-  const content = useMemo(() => {
-    // If comments are not enabled, return children as-is without wrapping in providers
-    if (!commentsEnabled) return children
-
-    return (
-      <CommentsProvider
-        documentId={documentId}
-        documentType={documentType}
-        isCommentsOpen={currentInspector?.name === COMMENTS_INSPECTOR_NAME}
-        onCommentsOpen={handleOpenCommentsInspector}
-      >
-        <CommentsSelectedPathProvider>{children}</CommentsSelectedPathProvider>
-      </CommentsProvider>
-    )
-  }, [
-    children,
-    commentsEnabled,
-    currentInspector?.name,
-    documentId,
-    documentType,
-    handleOpenCommentsInspector,
-  ])
-
   return (
-    <DocumentPaneContext.Provider value={documentPane}>
-      {inspectors.length > 0 && (
-        <DocumentInspectorMenuItemsResolver
-          documentId={documentId}
-          documentType={documentType}
-          inspectors={inspectors}
-          onMenuItems={setInspectorMenuItems}
-        />
-      )}
-
-      {/* Resolve root-level field actions */}
-      {fieldActions.length > 0 && schemaType && (
-        <FieldActionsResolver
-          actions={fieldActions}
-          documentId={documentId}
-          documentType={documentType}
-          onActions={setRootFieldActionNodes}
-          path={EMPTY_ARRAY}
-          schemaType={schemaType}
-        />
-      )}
-
-      <FieldActionsProvider actions={rootFieldActionNodes} path={EMPTY_ARRAY}>
-        {content}
-      </FieldActionsProvider>
-    </DocumentPaneContext.Provider>
+    <DocumentPaneContext.Provider value={documentPane}>{children}</DocumentPaneContext.Provider>
   )
 })
