+++ packages/sanity/src/desk/panes/document/document-layout/index.ts
@@ -1,0 +1,2 @@
+export * from './DocumentLayout'
+export * from './useDocumentLayoutComponent'
